% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarkeTest.R
\name{clarke_test}
\alias{clarke_test}
\title{Clarke Test}
\usage{
clarke_test(model1, model2, level=0.05, digits=2)
}
\arguments{
\item{model1}{A fitted statistical model of class \code{"game"},
\code{"lm"}, or \code{"glm"}}

\item{model2}{A fitted statistical model of class \code{"game"},
\code{"lm"}, or \code{"glm"} whose dependent variable is the same as that of
\code{model1}}

\item{level}{Numeric: significance level for the test.}

\item{digits}{Integer: number of digits to print}
}
\value{
Typical use will be to run the function interactively and examine
the printed output.  The functions return an object of class
\code{"nonnest.test"}, which is a list containing: \describe{
\item{\code{stat}}{The test statistic}
\item{\code{level}}{Significance level for the test}
\item{\code{digits}}{Number of digits to print}
\item{\code{loglik1}}{Vector of observationwise log-likelihoods for
\code{model1}}
\item{\code{loglik2}}{Vector of observationwise log-likelihoods for
\code{model2}}
\item{\code{nparams}}{Integer vector containing the number of parameters
fitted in \code{model1} and \code{model2} respectively}
\item{\code{nobs}}{Number of observations of the dependent variable being
modeled}}
}
\description{
`clarke_test` returns results from Kevin Clarke's distribution-free test
of non-nested models.
}
\details{
`clarke_test` is a more modularized version of the [clarke()] function from
the [games] package.  The main innovation is that the `nonnest` function
calls a generic `indivLogLiks` function, so additional methods can be easily
written for different classes of models.
}
\references{
Kevin Clarke.  2007.  "A Simple Distribution-Free Test for
Nonnested Hypotheses."  \emph{Political Analysis} 15(3): 347--363.
}
\author{
Brenton Kenkel (\email{brenton.kenkel@gmail.com}) modified by
Dave Armstrong (\email{dave@quantoid.net})
}
