% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarkeTest.R
\name{indivLogLiks}
\alias{indivLogLiks}
\alias{indivLogLiks.glm}
\alias{indivLogLiks.lm}
\title{Calculate individual log-likelihood values}
\usage{
indivLogLiks(model)

\method{indivLogLiks}{glm}(model)

\method{indivLogLiks}{lm}(model)
}
\arguments{
\item{model}{A statistical model object.}
}
\description{
Calculate individual log-likelihood values
}
