% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarkeTest.R
\name{nparams}
\alias{nparams}
\alias{nparams.glm}
\alias{nparams.lm}
\alias{nparams.polr}
\alias{nparams.clm}
\alias{nparams.multinom}
\alias{nparams.negbin}
\title{Find number of parameters in model}
\usage{
nparams(model)

\method{nparams}{glm}(model)

\method{nparams}{lm}(model)

\method{nparams}{polr}(model)

\method{nparams}{clm}(model)

\method{nparams}{multinom}(model)

\method{nparams}{negbin}(model)
}
\arguments{
\item{model}{A statistical model object.}
}
\value{
A scalar giving the number of parameters estimated in the model.
}
\description{
Find number of parameters in model
}
