% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarkeTest.R
\name{print.nonnest.test}
\alias{print.nonnest.test}
\title{Print non-nested test results}
\usage{
\method{print}{nonnest.test}(x, digits = x$digits, ...)
}
\arguments{
\item{x}{A result from the `nonnest` function}

\item{digits}{Number of digits to print in the output}

\item{...}{Other arguments passed down to print}
}
\value{
Printed output that summarises the results of the
\code{clarke_test} function.
}
\description{
Prints results of the \code{clarke_test} function.
}
