% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarkeTest.R
\name{indivLogLiks}
\alias{indivLogLiks}
\alias{indivLogLiks.glm}
\alias{indivLogLiks.lm}
\alias{indivLogLiks.polr}
\alias{indivLogLiks.clm}
\alias{indivLogLiks.multinom}
\alias{indivLogLiks.mlogit}
\alias{indivLogLiks.negbin}
\title{Calculate individual log-likelihood values}
\usage{
indivLogLiks(model)

\method{indivLogLiks}{glm}(model)

\method{indivLogLiks}{lm}(model)

\method{indivLogLiks}{polr}(model)

\method{indivLogLiks}{clm}(model)

\method{indivLogLiks}{multinom}(model)

\method{indivLogLiks}{mlogit}(model)

\method{indivLogLiks}{negbin}(model)
}
\arguments{
\item{model}{A statistical model object.}
}
\description{
Calculate individual log-likelihood values
}
\details{
The \code{indivLogLiks} function calls
the appropriate method for calculating individual
log likelihood values for the model.  The function
currently supports binomial, poisson and negative
binomial GLMs, ordinal models estimated with either
\code{\link{polr}} from the \code{\pkg{MASS}} package
or \code{\link{clm}} from the \code{\pkg{ordinal}}
package and multinomial models estimated with either
\code{\link{multinom}} from the \code{\pkg{nnet}}
package or \code{\link{mlogit}} from the
\code{\pkg{mlogit}} package.  Users can also write new
methods for both \code{indivLogLiks} and \code{nparams}
that would get called by the generic function.
}
