% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarkeTest.R
\name{indivLogLiks}
\alias{indivLogLiks}
\alias{indivLogLiks.glm}
\alias{indivLogLiks.lm}
\alias{indivLogLiks.polr}
\alias{indivLogLiks.clm}
\alias{indivLogLiks.multinom}
\alias{indivLogLiks.mlogit}
\alias{indivLogLiks.negbin}
\title{Calculate individual log-likelihood values}
\usage{
indivLogLiks(model)

\method{indivLogLiks}{glm}(model)

\method{indivLogLiks}{lm}(model)

\method{indivLogLiks}{polr}(model)

\method{indivLogLiks}{clm}(model)

\method{indivLogLiks}{multinom}(model)

\method{indivLogLiks}{mlogit}(model)

\method{indivLogLiks}{negbin}(model)
}
\arguments{
\item{model}{A statistical model object.}
}
\description{
Calculate individual log-likelihood values
}
