% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarkeTest.R
\name{nparams}
\alias{nparams}
\alias{nparams.glm}
\alias{nparams.lm}
\alias{nparams.polr}
\alias{nparams.clm}
\title{Find number of parameters in model}
\usage{
nparams(model)

\method{nparams}{glm}(model)

\method{nparams}{lm}(model)

\method{nparams}{polr}(model)

\method{nparams}{clm}(model)
}
\arguments{
\item{model}{A statistical model object.}
}
\description{
Find number of parameters in model
}
